+++ docs/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 ### Changed
 
+- Added extensible 2D base pass, changed Flat 2D to use the new system, and provided examples. ([#2451])
+
 ### Fixed
 
 ## [0.15.3] - 2020-08-22
@@ -33,6 +35,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 [#2447]: https://github.com/amethyst/amethyst/pull/2447
 [#2448]: https://github.com/amethyst/amethyst/pull/2448
+[#2451]: https://github.com/amethyst/amethyst/pull/2451
 
 ## [0.15.1] - 2020-08-14